+++ hmpc-rs/src/net/server.rs
@@ -93,9 +93,8 @@ async fn handle_connection(id: PartyID, #[cfg(feature = "sessions")] session: Se
 pub(crate) async fn run(id: PartyID, config: Config, data_channel: DataChannel, #[cfg(feature = "collective-consistency")] consistency_channel: ConsistencyCheckChannel)
 {
     let endpoint = make_server(id, &config).await;
-    // PANICS: Calling code checks that config.session is not None and contains a value
     #[cfg(feature = "sessions")]
-    let session = config.session.clone().unwrap().unwrap();
+    let session = config.session.clone();
     #[cfg(feature = "signing")]
     let verification_keys = Arc::new(config.verification_keys().await.expect("Could not get signature verification keys"));
     while let Some(connecting) = endpoint.accept().await